Extracts binary search and uses in seek as well
